Refactor SocialMediaButton to SocialMediaLink and get rid of IconButton of material-ui
